requirements.txt
click
Biopython
